Extract Discord response facade from service
